interface.c: split out free_prefix_list
The prefix list from the interface is going to be used to generate
a new list of prefixes to send in an advert.  As such, it makes
since to reuse a common free_prefix_list function to clean up both
the iface->prefix list and the prefix list generated in send.c.