device-linux.c: fix check_ip6_forwarding retval
This function was attempting to log a failure only once, but was
returning the wrong value in all but the first call.  This affected
the values set in build_ra_header.  This commit ensures the return
value is correct, and changes the log level from 1 to 5 after the
first call.